feat: easier navigation for review comments
